arreglat problemas amb la barra de navegacio

+++ vista/foro.view.php
@@ -11,7 +11,6 @@
     <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"></script>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
 
-    <script type="module" src="../Personatje/buscarPersonatges.js"></script>
     <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"></script>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
@@ -22,70 +21,65 @@
         <h1>Foro de Discusión</h1>
     </header>
 
-              
     <nav class="navbar navbar-expand-lg navbar-dark navegacio">
-        <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
-                <li class="nav-item">
-                    <a class="Inici nav-link " href="../controlador/index.controler.php ">FinalD</a>
-
-                </li>
-                <li class="nav-item">
-                    <a href="../controlador/personatges.controller.php" class="Personatges  nav-link ">Personatges </a>
-
-                </li>
-                <li class="nav-item">
-                    <a href="../controlador/foro.controller.php" class="nav-link active">Forum </a>
-                    
-                </li>
-                
-                <li class="nav-item">
-                    <a class="nav-link" href="../vista/crearForo.view.php">Crear nou Missatge</a>
-                </li>
-
-                <li class="nav-item">
-                    <a class="nav-link" href="../controlador/articlesPropis.controller.php">Els teus articles</a>
-                </li>
-
-            </ul>
-
+    <div class="container-fluid">  
+        <a class="Inici nav-link " href="../controlador/index.controler.php ">FinalD</a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                 <span class="navbar-toggler-icon"></span>
             </button>
-            <?php
-            if(isset($_SESSION['user'])){
-                echo '<div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav
-                ml-auto">
-                <li class="nav-item">
-                <div id="name"  class="nav-link" ">'.$_SESSION['username'].'</div>
-                </li>
-                <li class="nav-item">
-                <a class="nav-link" href="../controlador/logout.controller.php">Logout</a>
-                </li>
-                </ul>
-                </div>';
-            }else{
-                echo '<div class="collapse navbar-collapse" id="navbarNav">
-                <ul class="navbar-nav
-                ml-auto">
-                <li class="nav-item">
-                <a class="nav-link" href="../vista/login.php">Login</a>
-                </li>
-                <li class="nav-item">
-                <a class="nav-link" href="../vista/registre.php">Register</a>
+            <div class="collapse navbar-collapse" id="navbarNav">
+                <ul class="navbar-nav ">
+                    <li class="nav-item">
+                        <a href="../controlador/personatges.controller.php" class="Personatges  nav-link">Personatges </a>
+                        
+                    </li>
+                    <li class="nav-item dropdown">
+                        <a class="nav-link active dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                         Forum
+                        </a>
+                        <div class="dropdown-menu" aria-labelledby="navbarDropdown">
+                            <a class="dropdown-item" href="../controlador/foro.controller.php">Inici</a>
+                            <div class="dropdown-divider"></div>
+                            <a class="dropdown-item" href="../controlador/crearForo.controller.php">Crear missatge</a>
+                            <a class="dropdown-item" href="../controlador/articlesPropis.controller.php">Els teus articles</a>
+                        </div>
+                    </li>            
+                    
+                </ul>    
                 
-                </li>
-                </ul>
-                </div>';
-            }
+                 
+                <?php
+                if(isset($_SESSION['user'])){
+                    echo '<div class="collapse navbar-collapse" id="navbarNav">
+                    <ul class="navbar-nav ml-auto">
+                    <li class="nav-item">
+                    <div id="name"  class="nav-link" ">'.$_SESSION['username'].'</div>
+                    </li>
+                    <li class="nav-item">
+                    <a class="nav-link" href="../controlador/logout.controller.php">Logout</a>
+                    </li>
+                    </ul>
+                    </div>';
+                }else{
+                    echo '<div class="collapse navbar-collapse" id="navbarNav">
+                    <ul class="navbar-nav ml-auto">
+                    <li class="nav-item">
+                    <a class="nav-link" href="../vista/login.php">Login</a>
+                    </li>
+                    <li class="nav-item">
+                    <a class="nav-link" href="../vista/registre.php">Register</a>
+                    
+                    </li>
+                    </ul>
+                    </div>';
+                }
+                
+                ?>
+               
             
-            ?>
-        
+            </div>
         </div>
     </nav>
-    
-
 
     <?php
 
@@ -115,9 +109,14 @@
                     <p class="card-text">'.$row['missatge'].'</p>
                     <small class="text-muted">Publicado por: '.$row['nom'].' | Fecha: '.$row['data'].'</small>';
                     
-                    if($row['id_Usuari'] == $_SESSION['user']){
-                        echo '<a href="../controlador/eliminarArticle.controller.php?id='.$row['id'].'" class="btn btn-danger delete">Eliminar</a>';
-                    };
+
+                    if(isset($_SESSION['user'])){
+                        
+                        
+                        if($row['id_Usuari'] == $_SESSION['user']){
+                            echo '<a href="../controlador/eliminarArticle.controller.php?id='.$row['id'].'" class="btn btn-danger delete">Eliminar</a>';
+                        };
+                    }
 
                     echo '</div></div>';
             }
@@ -139,10 +138,7 @@
 
     </footer>
 
-    
-    <!-- Scripts de Bootstrap y jQuery (asegúrate de incluir jQuery antes de Bootstrap) -->
-    <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js"></script>
-    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+
+   
 </body>
 </html>
